HomeViewModal has more responsibilities adding & removing article label
